+++ tools/yabridgectl/src/utils.rs
@@ -316,18 +316,23 @@ pub fn verify_wine_setup(config: &mut Config) -> Result<()> {
         eprintln!(
             "\n{}",
             wrap(&format!(
-                "Warning: Could not run 'yabridge-host.exe'. Wine reported the following error: \n\
+                "Warning: Could not run '{yabridge_host}'. Wine reported the following error:\n\
                  \n\
-                 {}\n\
+                 {error}\n\
                  \n\
-                 This can happen when using a version of Wine that is much older than the version \
-                 that has been used to compile yabridge with. Your current Wine version is '{}'. \
-                 See the troubleshooting section of the readme for more information on how to \
-                 upgrade your installation of Wine.\n\
+                 Make sure that you have downloaded the correct version of yabridge for your distro.\n\
+                 This can also happen when using a version of Wine that's not compatible with this \
+                 version of yabridge, in which case you'll need to upgrade Wine. Your current Wine \
+                 version is '{wine_version}'. \
+                 See the link below for instructions on how to upgrade your installation of Wine.\n\
+                 Finally, if the default Wine prefix over at {wine_prefix} is set to 32-bit only, \
+                 this check will also fail as the 64-bit {yabridge_host} won't be able to run.
                  \n\
                  https://github.com/robbert-vdh/yabridge#troubleshooting-common-issues",
-                last_error.unwrap_or("<no_output>").bright_white(),
-                wine_version
+                yabridge_host = "yabridge-host.exe".bright_white(),
+                wine_prefix = "~/.wine".bright_white(),
+                error = last_error.unwrap_or("<no_output>").bright_white(),
+                wine_version = wine_version
                     .strip_prefix("wine-")
                     .unwrap_or(&wine_version)
                     .bright_white(),